fix: use uint256 for getLastUpdateTimestamp ABI to prevent protocol data crash
ethers.js v5 decodes uint40 as a plain JS number, not BigNumber.
The fetchProtocolData() call to lastTimestamp.toNumber() crashed with
TypeError, preventing all protocol stats (Total Staked, Net APY) from
loading on the Yield page. Changing the ABI return type to uint256
makes ethers.js return BigNumber consistently.

+++ src/abis/StakedXCN.ts
@@ -71,7 +71,7 @@ export const STAKED_XCN_ABI = [
   {
     inputs: [],
     name: 'getLastUpdateTimestamp',
-    outputs: [{ name: '', type: 'uint40' }],
+    outputs: [{ name: '', type: 'uint256' }],
     stateMutability: 'view',
     type: 'function',
   },
